[DocsUI tests] Consistently use Long for timeout in bots
The timeout should be a Long but we have defined it as an Int and
relied on the auto-conversion to Long. Let's define it as a Long
from the get-go.

Bug: 483492490
Test: EXEMPT PURE_REFACTOR
Flag: EXEMPT PURE_REFACTOR

+++ tests/common/com/android/documentsui/bots/DirectoryListBot.java
@@ -142,7 +142,7 @@ public class DirectoryListBot extends Bots.BaseBot {
             UiDevice device,
             UiAutomation automation,
             Context context,
-            int timeout,
+            long timeout,
             @LayoutRes Integer layoutId) {
         super(device, context, timeout, layoutId);
         mAutomation = automation;
@@ -636,14 +636,14 @@ public class DirectoryListBot extends Bots.BaseBot {
     }
 
     public void assertOrder(String[] dirs, String[] files) throws UiObjectNotFoundException {
-        int remaining = mTimeout;
+        long remaining = mTimeout;
         if (remaining < 0) {
             remaining = 0;
         }
         // 1048576 is (1 << 20), a power of two close to one million. The value is basically
         // arbitrary. We just want our sleeps to start as a small fraction of the default timeout,
         // but double in length each iteration.
-        int retryTimeout = remaining / 1048576;
+        long retryTimeout = remaining / 1048576;
         if ((retryTimeout < 1) && (remaining != 0)) {
             retryTimeout = 1;
         }
